requirements.txt
streamlit==1.35.0
fastapi==0.111.0
uvicorn==0.30.1
langchain==0.3.0
langchain-community==0.3.0
faiss-cpu==1.8.0
sentence-transformers>=3.0.0
pypdf2==3.0.1
python-dotenv==1.0.1
pydantic==2.7.4